Split config into slumber_config crate
This solves the fuckery around having multiple config structs.
