+++ Cargo.toml
@@ -12,18 +12,20 @@ version = "1.7.0"
 rust-version = "1.80.0"
 
 [workspace]
-members = ["crates/slumber_cli", "crates/slumber_core", "crates/slumber_tui"]
+members = ["crates/slumber_cli", "crates/slumber_config", "crates/slumber_core", "crates/slumber_tui"]
 
 # Dependencies used in multiple crates get de-duplicated here
 [workspace.dependencies]
 anyhow = "1.0.0"
 bytes = {version = "1.6.1", default-features = false}
 chrono = {version = "0.4.31", default-features = false}
+crossterm = {version = "0.27.0", default-features = false}
 derive_more = {version = "1.0.0-beta.6", default-features = false}
 futures = "0.3.28"
 indexmap = {version = "2.0.0", default-features = false}
 itertools = "0.13.0"
 pretty_assertions = "1.4.0"
+ratatui = {version = "0.27.0", default-features = false}
 reqwest = {version = "0.12.5", default-features = false}
 rstest = {version = "0.21.0", default-features = false}
 serde = {version = "1.0.204", default-features = false}
@@ -31,6 +33,10 @@ serde_json = {version = "1.0.120", default-features = false}
 serde_json_path = "0.6.3"
 serde_test = "1.0.176"
 serde_yaml = {version = "0.9.0", default-features = false}
+slumber_cli = {path = "./crates/slumber_cli", version = "1.7.0"}
+slumber_config = {path = "./crates/slumber_config", version = "1.7.0"}
+slumber_core = {path = "./crates/slumber_core", version = "1.7.0"}
+slumber_tui = {path = "./crates/slumber_tui", version = "1.7.0"}
 strum = {version = "0.26.3", default-features = false}
 thiserror = "1.0.63"
 tokio = {version = "1.39.2", default-features = false}
@@ -39,9 +45,9 @@ uuid = {version = "1.10.0", default-features = false}
 
 [dependencies]
 anyhow = {workspace = true, features = ["backtrace"]}
-slumber_cli = {path = "./crates/slumber_cli", version = "1.7.0"}
-slumber_core = {path = "./crates/slumber_core", version = "1.7.0"}
-slumber_tui = {path = "./crates/slumber_tui", version = "1.7.0"}
+slumber_cli = {workspace = true}
+slumber_core = {workspace = true}
+slumber_tui = {workspace = true}
 tokio = {workspace = true, features = ["macros", "rt-multi-thread"]}
 tracing = {workspace = true}
 tracing-subscriber = {version = "0.3.17", default-features = false, features = ["ansi", "fmt", "registry"]}
