+++ crates/slumber_cli/src/commands/show.rs
@@ -1,9 +1,9 @@
 use crate::{GlobalArgs, Subcommand};
 use clap::Parser;
 use serde::Serialize;
+use slumber_config::Config;
 use slumber_core::{
     collection::CollectionFile,
-    config::{self, Config},
     db::Database,
     util::{DataDirectory, TempDirectory},
 };
@@ -34,7 +34,7 @@ impl Subcommand for ShowCommand {
                     CollectionFile::try_path(None, global.file);
                 println!("Data directory: {}", DataDirectory::get());
                 println!("Temporary directory: {}", TempDirectory::get());
-                println!("Config: {}", config::path().display());
+                println!("Config: {}", Config::path().display());
                 println!("Database: {}", Database::path().display());
                 println!(
                     "Collection: {}",
