+++ crates/slumber_tui/src/view/event.rs
@@ -1,14 +1,12 @@
 //! Utilities for handling input events from users, as well as external async
 //! events (e.g. HTTP responses)
 
-use crate::{
-    input::Action,
-    view::{
-        common::modal::{Modal, ModalPriority},
-        state::{Notification, RequestState},
-        Component,
-    },
+use crate::view::{
+    common::modal::{Modal, ModalPriority},
+    state::{Notification, RequestState},
+    Component,
 };
+use slumber_config::Action;
 use slumber_core::http::RequestId;
 use std::{any::Any, collections::VecDeque, fmt::Debug, ops::DerefMut};
 use tracing::trace;
